@@ -1,7 +1,7 @@
 package io.virtualan.message.core.jms;
 
-import javax.jms.ConnectionFactory;
-import javax.jms.JMSException;
+import jakarta.jms.ConnectionFactory;
+import jakarta.jms.JMSException;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnClass;
 import org.springframework.stereotype.Service;
 
@@ -10,7 +10,7 @@ import org.springframework.stereotype.Service;
 public class ActiveMQConnectionFactory implements VirtualanJMSConnectionFactory  {
 
   @Override
-  public ConnectionFactory connectionFactory(JMSConfigurationDomain conf) throws JMSException {
+  public jakarta.jms.ConnectionFactory connectionFactory(JMSConfigurationDomain conf) throws JMSException {
     ConnectionFactory connectionFactory = null;
     if (conf.getUserName() != null && conf.getPassword() != null) {
       connectionFactory = new org.apache.activemq.ActiveMQConnectionFactory(conf.getUserName(), conf.getPassword(),
