@@ -1,8 +1,9 @@
 package io.virtualan.message.core.jms;
 
-import com.ibm.mq.jms.MQQueueConnectionFactory;
-import javax.jms.ConnectionFactory;
-import javax.jms.JMSException;
+
+import com.ibm.mq.jakarta.jms.MQQueueConnectionFactory;
+import jakarta.jms.ConnectionFactory;
+import jakarta.jms.JMSException;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnClass;
 import org.springframework.stereotype.Service;
 
