@@ -17,6 +17,7 @@ package io.virtualan.core.model;
 import java.util.List;
 import java.util.Set;
 import lombok.Data;
+import lombok.Getter;
 import lombok.Setter;
 
 
@@ -27,7 +28,7 @@ import lombok.Setter;
  * @author  Elan Thangamani
  * 
  **/
-@Data
+
 public class MockRequest {
     private long virtualServiceId;
     private long usageCount;
@@ -64,5 +65,93 @@ public class MockRequest {
             ", availableParams=" + availableParams +
             ", headerParams=" + headerParams +
             '}';
+    }
+
+    public long getVirtualServiceId() {
+        return virtualServiceId;
+    }
+
+    public void setVirtualServiceId(long virtualServiceId) {
+        this.virtualServiceId = virtualServiceId;
+    }
+
+    public long getUsageCount() {
+        return usageCount;
+    }
+
+    public void setUsageCount(long usageCount) {
+        this.usageCount = usageCount;
+    }
+
+    public String getInput() {
+        return input;
+    }
+
+    public void setInput(String input) {
+        this.input = input;
+    }
+
+    public String getRule() {
+        return rule;
+    }
+
+    public void setRule(String rule) {
+        this.rule = rule;
+    }
+
+    public String getType() {
+        return type;
+    }
+
+    public void setType(String type) {
+        this.type = type;
+    }
+
+    public ResponseProcessType getResponseProcessType() {
+        return responseProcessType;
+    }
+
+    public void setResponseProcessType(ResponseProcessType responseProcessType) {
+        this.responseProcessType = responseProcessType;
+    }
+
+    public Set getExcludeSet() {
+        return excludeSet;
+    }
+
+    public void setExcludeSet(Set excludeSet) {
+        this.excludeSet = excludeSet;
+    }
+
+    public String getMethod() {
+        return method;
+    }
+
+    public void setMethod(String method) {
+        this.method = method;
+    }
+
+    public ContentType getContentType() {
+        return contentType;
+    }
+
+    public void setContentType(ContentType contentType) {
+        this.contentType = contentType;
+    }
+
+    public List<VirtualServiceKeyValue> getAvailableParams() {
+        return availableParams;
+    }
+
+    public void setAvailableParams(List<VirtualServiceKeyValue> availableParams) {
+        this.availableParams = availableParams;
+    }
+
+    public List<VirtualServiceKeyValue> getHeaderParams() {
+        return headerParams;
+    }
+
+    public void setHeaderParams(List<VirtualServiceKeyValue> headerParams) {
+        this.headerParams = headerParams;
     }
 }
